+++ hcore/hset.h
@@ -63,7 +63,7 @@ private:
 	HSBBSTree f_oEngine;
 public:
 	HSet( void ) : f_oEngine() {};
-	size_t size( void ) const
+	int long size( void ) const
 		{ return ( f_oEngine.size() );	}
 	bool empty( void ) const
 		{ return ( f_oEngine.empty() );	}
